+++ qiskit_aqua_ui/run/_controller.py
@@ -17,28 +17,24 @@
 
 from ._model import Model
 from ._customwidgets import (EntryPopup, ComboboxPopup, TextPopup)
-import psutil
 import os
-import subprocess
 import threading
 import queue
-import tempfile
 import tkinter as tk
 from tkinter import messagebox
 import json
 import ast
-import sys
 import logging
+from qiskit_aqua_ui import GUIProvider
 
 logger = logging.getLogger(__name__)
 
 
 class Controller(object):
 
-    _START, _STOP = 'Start', 'Stop'
-
-    def __init__(self):
+    def __init__(self, guiprovider):
         self._view = None
+        self._guiprovider = guiprovider
         self._model = Model()
         self._filemenu = None
         self._title = tk.StringVar()
@@ -53,7 +49,7 @@ class Controller(object):
         self._start_button = None
         self._thread_queue = queue.Queue()
         self._thread = None
-        self._command = Controller._START
+        self._command = GUIProvider.START
         self._process_stop = False
         self._validate_integer_command = None
         self._validate_float_command = None
@@ -374,10 +370,8 @@ class Controller(object):
     def on_section_property_remove(self, section_name, property_name):
         try:
             self.model.delete_section_property(section_name, property_name)
-            self._propertiesView.populate(
-                self.model.get_section_properties(section_name))
-            self._propertiesView.show_add_button(
-                self.shows_add_button(section_name))
+            self._propertiesView.populate(self.model.get_section_properties(section_name))
+            self._propertiesView.show_add_button(self.shows_add_button(section_name))
             self._propertiesView.show_remove_button(False)
             self._propertiesView.show_defaults_button(not self.model.default_properties_equals_properties(section_name))
         except Exception as e:
@@ -478,12 +472,18 @@ class Controller(object):
         self._filemenu.entryconfig(2, state='disabled')
         self._view.after(100, self._process_thread_queue)
         try:
-            if self._command is Controller._START:
+            if self._command is GUIProvider.START:
                 self._outputView.clear()
-                self._thread = AlgorithmThread(
-                    self.model, self._outputView, self._thread_queue)
-                self._thread.daemon = True
-                self._thread.start()
+                self._thread = self._guiprovider.create_run_thread(self.model, self._outputView, self._thread_queue)
+                if self._thread is not None:
+                    self._thread.daemon = True
+                    self._thread.start()
+                else:
+                    self._thread_queue.put(None)
+                    self._start_button.state(['!disabled'])
+                    self._filemenu.entryconfig(0, state='normal')
+                    self._filemenu.entryconfig(1, state='normal')
+                    self._filemenu.entryconfig(2, state='normal')
             else:
                 self.stop()
         except Exception as e:
@@ -499,13 +499,13 @@ class Controller(object):
         if self._thread is not None:
             stopthread = threading.Thread(target=Controller._stop,
                                           args=(self._thread,),
-                                          name='Aqua stop thread')
+                                          name='Stop thread')
             stopthread.daemon = True
             stopthread.start()
             self._outputView.clear_buffer()
             self._thread = None
             self._process_stop = True
-            self._thread_queue.put(Controller._STOP)
+            self._thread_queue.put(GUIProvider.STOP)
 
     @staticmethod
     def _stop(thread):
@@ -520,19 +520,19 @@ class Controller(object):
             line = self._thread_queue.get_nowait()
             if line is None:
                 return
-            elif line is Controller._START:
+            elif line is GUIProvider.START:
                 self._progress.start(500)
-                self._command = Controller._STOP
+                self._command = GUIProvider.STOP
                 self._button_text.set(self._command)
                 self._start_button.state(['!disabled'])
-            elif line is Controller._STOP:
+            elif line is GUIProvider.STOP:
                 if not self._outputView.buffer_empty():
                     # repost stop
-                    self._thread_queue.put(Controller._STOP)
+                    self._thread_queue.put(GUIProvider.STOP)
                 else:
                     self._thread = None
                     self._progress.stop()
-                    self._command = Controller._START
+                    self._command = GUIProvider.START
                     self._button_text.set(self._command)
                     self._start_button.state(['!disabled'])
                     self._filemenu.entryconfig(0, state='normal')
@@ -548,107 +548,3 @@ class Controller(object):
             pass
 
         self._view.after(100, self._process_thread_queue)
-
-
-class AlgorithmThread(threading.Thread):
-
-    def __init__(self, model, output, queue):
-        super(AlgorithmThread, self).__init__(name='Algorithm run thread')
-        self.model = model
-        self._output = output
-        self._thread_queue = queue
-        self._popen = None
-
-    def stop(self):
-        self._output = None
-        self._thread_queue = None
-        if self._popen is not None:
-            p = self._popen
-            self._kill(p.pid)
-            p.stdout.close()
-
-    def _kill(self, proc_pid):
-        try:
-            process = psutil.Process(proc_pid)
-            for proc in process.children(recursive=True):
-                proc.kill()
-            process.kill()
-        except Exception as e:
-            if self._output is not None:
-                self._output.write_line(
-                    'Process kill has failed: {}'.format(str(e)))
-
-    def run(self):
-        input_file = None
-        temp_input = False
-        try:
-            algorithms_directory = os.path.dirname(os.path.realpath(__file__))
-            algorithms_directory = os.path.abspath(
-                os.path.join(algorithms_directory, '../../qiskit_aqua_cmd'))
-            input_file = self.model.get_filename()
-            if input_file is None or self.model.is_modified():
-                fd, input_file = tempfile.mkstemp(suffix='.in')
-                os.close(fd)
-                temp_input = True
-                self.model.save_to_file(input_file)
-
-            startupinfo = None
-            process_name = psutil.Process().exe()
-            if process_name is None or len(process_name) == 0:
-                process_name = 'python'
-            else:
-                if sys.platform == 'win32' and process_name.endswith('pythonw.exe'):
-                    path = os.path.dirname(process_name)
-                    files = [f for f in os.listdir(path) if f != 'pythonw.exe' and f.startswith(
-                        'python') and f.endswith('.exe')]
-                    # sort reverse to have hihre python versions first: python3.exe before python2.exe
-                    files = sorted(files, key=str.lower, reverse=True)
-                    new_process = None
-                    for file in files:
-                        p = os.path.join(path, file)
-                        if os.path.isfile(p):
-                            # python.exe takes precedence
-                            if file.lower() == 'python.exe':
-                                new_process = p
-                                break
-
-                            # use first found
-                            if new_process is None:
-                                new_process = p
-
-                    if new_process is not None:
-                        startupinfo = subprocess.STARTUPINFO()
-                        startupinfo.dwFlags = subprocess.STARTF_USESHOWWINDOW
-                        startupinfo.wShowWindow = subprocess.SW_HIDE
-                        process_name = new_process
-
-            if self._output is not None and logger.getEffectiveLevel() == logging.DEBUG:
-                self._output.write('Process: {}\n'.format(process_name))
-
-            self._popen = subprocess.Popen([process_name,
-                                            algorithms_directory,
-                                            input_file],
-                                           stdin=subprocess.DEVNULL,
-                                           stdout=subprocess.PIPE,
-                                           stderr=subprocess.STDOUT,
-                                           universal_newlines=True,
-                                           startupinfo=startupinfo)
-            if self._thread_queue is not None:
-                self._thread_queue.put(Controller._START)
-            for line in iter(self._popen.stdout.readline, ''):
-                if self._output is not None:
-                    self._output.write(str(line))
-            self._popen.stdout.close()
-            self._popen.wait()
-        except Exception as e:
-            if self._output is not None:
-                self._output.write('Process has failed: {}'.format(str(e)))
-        finally:
-            self._popen = None
-            if self._thread_queue is not None:
-                self._thread_queue.put(Controller._STOP)
-
-            if temp_input and input_file is not None:
-                os.remove(input_file)
-
-            input_file = None
